+++ src/components/pro/RTablePro/index.ts
@@ -3,9 +3,23 @@ import tableProProps from './src/props'
 import { useTablePro } from './src/hooks/useTablePro'
 
 import type { ExtractPropTypes } from 'vue'
+import type {
+  BasePagination,
+  TablePagination,
+  FormatRangeTime,
+  TableRequestConfig,
+  TableProFieldNames,
+} from './src/types'
 
 type TableProProps = ExtractPropTypes<typeof tableProProps>
 
-export type { TableProProps }
+export type {
+  TableProProps,
+  BasePagination,
+  TablePagination,
+  FormatRangeTime,
+  TableRequestConfig,
+  TableProFieldNames,
+}
 
 export { RTablePro, useTablePro, tableProProps }