+++ src/views/demo/form/index.tsx
@@ -9,11 +9,18 @@ import {
   NButton,
   NRadio,
   NRadioGroup,
+  NCard,
 } from 'naive-ui'
 
-import { useForm, useModal } from '@/components'
+import { useForm } from '@/components'
 
-import type { RFormRules } from '@/components'
+interface FormModel {
+  name: string | null
+  age: number | null
+  gender: string | null
+  date: Date | null
+  remark: string | null
+}
 
 export default defineComponent({
   name: 'RFormDemo',
@@ -21,8 +28,15 @@ export default defineComponent({
     // 使用以下 hooks 的时候，应该注意调用时机
     const [
       register,
-      { getFormInstance, validate, restoreValidation, formModel, formRules },
-    ] = useForm(
+      {
+        getFormInstance,
+        validate,
+        restoreValidation,
+        formModel,
+        formRules,
+        reset,
+      },
+    ] = useForm<FormModel>(
       {
         name: null,
         age: null,
@@ -35,6 +49,7 @@ export default defineComponent({
           required: true,
           message: '请输入姓名',
           trigger: ['blur', 'change'],
+          key: 'name',
         },
         date: {
           required: true,
@@ -59,90 +74,99 @@ export default defineComponent({
     /**
      *
      * @description
-     * 如果待验证数据类型为: number, array 等，需要手动设置 type 类型。
-     * 具体可以吃查看: async-validator type
-     * @see https://github.com/yiminghe/async-validator?tab=readme-ov-file#type
-     *
-     * 如果你需要自定义验证，可以查看：naive ui custom validation
-     * @see https://www.naiveui.com/zh-CN/dark/components/form#custom-validation.vue
-     *
-     * 如果只是简单的 rules 管理，可以在初始化 useForm 的时候传入第二个参数；
-     * 然后使用 formRules 方法获取到初始化 rules 数据。
-     */
-    const rules = ref(formRules())
-    /**
-     *
-     * @description
      * 如果只是简单的数据，可以在初始化 useForm 的时候直接传入第一个参数；
      * 然后使用 formModel 方法获取到初始化 model 数据。
      *
-     * 动态的复杂数据，不建议使用该方法管理 model；手动的拆分出来是一个更加好的选择。
+     * 当然，在 vue 中，推荐使用该方法进行表单值的初始化操作；
+     * 因为，可以在你需要的时候实现便捷的一键初始化表单需求。
      */
     const condition = ref(formModel())
 
+    /**
+     *
+     * @param key 需要校验的字段。
+     *
+     * @see https://www.naiveui.com/zh-CN/dark/components/form#partially-apply-rules.vue
+     *
+     * @description
+     * 执行部分的校验规则。
+     */
+    const onlyValidateSection = (key: string) => {
+      validate(void 0, (rules) => {
+        return rules?.key === key
+      })
+    }
+
     return {
       register,
-      rules,
       condition,
       restoreValidation,
-      formModel,
       validate,
+      formRules,
+      onlyValidateSection,
+      reset,
     }
   },
   render() {
-    const { rules } = this
-    const { register, restoreValidation, formModel, validate } = this
+    const { condition } = this
+    const {
+      register,
+      restoreValidation,
+      validate,
+      formRules,
+      onlyValidateSection,
+      reset,
+    } = this
 
     return (
-      <RForm onRegister={register} rules={rules} model={this.condition}>
-        <NGrid cols={24} xGap={24}>
-          <NFormItemGi label="姓名" path="name" span={12}>
-            <NInput v-model:value={this.condition.name} />
-          </NFormItemGi>
-          <NFormItemGi label="年龄" path="age" span={12}>
-            <NInputNumber
-              v-model:value={this.condition.age}
-              showButton={false}
-              style="width: 100%"
-            />
-          </NFormItemGi>
-          <NFormItemGi label="出生日期" path="date" span={12}>
-            <NDatePicker
-              v-model:value={this.condition.date}
-              style="width: 100%"
-            />
-          </NFormItemGi>
-          <NFormItemGi label="性别" path="gender" span={12}>
-            <NRadioGroup v-model:value={this.condition.gender}>
-              <NRadio value="girl">女</NRadio>
-              <NRadio value="man">男</NRadio>
-            </NRadioGroup>
-          </NFormItemGi>
-          <NFormItemGi label="备注信息" span={24}>
-            <NInput type="textarea" v-model:value={this.condition.remark} />
-          </NFormItemGi>
-          <NFormItemGi span={24}>
-            <NFlex justify="flex-end" style="width: 100%">
-              <NButton
-                type="info"
-                onClick={() => {
-                  this.condition = formModel()
-
-                  restoreValidation()
-                }}
-              >
-                重置表单为初始状态
-              </NButton>
-              <NButton type="warning" onClick={restoreValidation.bind(this)}>
-                移除校验状态
-              </NButton>
-              <NButton type="primary" onClick={() => validate()}>
-                校验
-              </NButton>
-            </NFlex>
-          </NFormItemGi>
-        </NGrid>
-      </RForm>
+      <NCard title="useForm 表单校验">
+        <RForm onRegister={register} rules={formRules()} model={condition}>
+          <NGrid cols={24} xGap={24}>
+            <NFormItemGi label="姓名" path="name" span={12}>
+              <NInput v-model:value={condition.name} />
+            </NFormItemGi>
+            <NFormItemGi label="年龄" path="age" span={12}>
+              <NInputNumber
+                v-model:value={condition.age}
+                showButton={false}
+                style="width: 100%"
+              />
+            </NFormItemGi>
+            <NFormItemGi label="出生日期" path="date" span={12}>
+              <NDatePicker v-model:value={condition.date} style="width: 100%" />
+            </NFormItemGi>
+            <NFormItemGi label="性别" path="gender" span={12}>
+              <NRadioGroup v-model:value={condition.gender}>
+                <NRadio value="girl">女</NRadio>
+                <NRadio value="man">男</NRadio>
+              </NRadioGroup>
+            </NFormItemGi>
+            <NFormItemGi label="备注信息" span={24}>
+              <NInput type="textarea" v-model:value={condition.remark} />
+            </NFormItemGi>
+            <NFormItemGi span={24}>
+              <NFlex justify="flex-end" style="width: 100%">
+                <NButton type="info" onClick={() => reset(this.condition)}>
+                  重置表单为初始状态
+                </NButton>
+                <NButton type="warning" onClick={restoreValidation.bind(this)}>
+                  移除校验状态
+                </NButton>
+                <NButton
+                  type="primary"
+                  onClick={() => onlyValidateSection('name')}
+                >
+                  仅校验姓名字段
+                </NButton>
+                <NButton type="primary" onClick={() => validate()}>
+                  校验
+                </NButton>
+                <NButton attrType="reset">重置</NButton>
+              </NFlex>
+            </NFormItemGi>
+          </NGrid>
+        </RForm>
+      </NCard>
     )
   },
 })